+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+### Changed
+- Update overall docstrings
+- Opt out of typing module when deprecated
+- Refactor kaioretry.retry and kaioretry.aioretry through another
+  decorator.
+- Expose logger parameter in kaioretry.retry and kaioretry.aioretry
+  signatures.
+- Fix return description in kaioretry.retry and kaioretry.aioretry
+  docstrings.
+
 ## [0.3.0] - 2023-01-29
 
 ### Fixed